+++ CHANGES.rst
@@ -38,6 +38,12 @@ Changelog
   delay aware adviser config if it was never used and to change the
   `for_item_created_from` if it is not an auto asked advice.
   [gbastien]
+- Cleaned `UnrestrictedMethodsView`, splitted it to `ItemUnrestrictedMethodsView`
+  and `MeetingUnrestrictedMethodsView` because the `findFirstItemNumberForMeeting`
+  method is the only one called with a `Meeting` as context and others need a
+  `MeetingItem` as context.
+  Renamed `findFirstItemNumberForMeeting` to `findFirstItemNumber`.
+  [gbastien]
 
 4.2rc29 (2022-06-17)
 --------------------